--- expect/_to_strict_equal.ts
@@ -1,20 +1,0 @@
-// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-
-import { MatcherContext, MatchResult } from "./_types.ts";
-import { assertNotStrictEquals } from "../assert/assert_not_strict_equals.ts";
-import { assertStrictEquals } from "../assert/assert_strict_equals.ts";
-
-/* Similar to assertStrictEquals */
-export function toStrictEqual(
-  context: MatcherContext,
-  expected: unknown,
-): MatchResult {
-  if (context.isNot) {
-    return assertNotStrictEquals(
-      context.value,
-      expected,
-      context.customMessage,
-    );
-  }
-  return assertStrictEquals(context.value, expected, context.customMessage);
-}
